Updated version fixes

+++ common/spl/spl.c
@@ -172,7 +172,7 @@ int spl_parse_image_header(struct spl_image_info *spl_image,
 		spl_image->name = image_get_name(header);
 		debug("spl: payload image: %.*s load addr: 0x%lx size: %d\n",
 			IH_NMLEN, spl_image->name,
-			spl_image->load_addr, spl_image->size);
+			(unsigned long int) spl_image->load_addr, (u32)spl_image->size);
 #else
 		/* LEGACY image not supported */
 		debug("Legacy boot image support not enabled, proceeding to other boot methods\n");
@@ -228,7 +228,7 @@ __weak void __noreturn jump_to_image_no_args(struct spl_image_info *spl_image)
 	image_entry_noargs_t image_entry =
 		(image_entry_noargs_t)spl_image->entry_point;
 
-	debug("image entry point: 0x%lX\n", spl_image->entry_point);
+	debug("image entry point: 0x%lX\n", (unsigned long int)spl_image->entry_point);
 	image_entry();
 }
